api/app/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
fastapi==0.63.0
iso8601==0.1.13
passlib==1.7.4
priority==1.3.0
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
PyPika==0.44.1
python-multipart==0.0.5
pytz==2020.5
six==1.15.0
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.19
typing-extensions==3.7.4.3
wsproto==1.0.0
asyncpg
uvicorn
anyio
requests
python-dotenv
fastapi-mail